+++ src/formatters/ios_formatters.ts
@@ -1,10 +1,15 @@
 import {FormatterArguments} from 'style-dictionary/types/Format';
-import {_themeColorTokens} from './common';
+import {_themeColorTokens, _themeTypographyTokens} from './common';
 import {File} from 'style-dictionary';
 import * as StyleDictionary from 'style-dictionary';
 import camelCase from 'camelcase';
 import {compileTemplate} from '../utils/utils';
+import {
+  transformFontWeight,
+  transformPercentValue,
+} from '../utils/transform_utils';
 import Handlebars = require('handlebars');
+import {snakeCase} from 'snake-case';
 
 const {fileHeader} = StyleDictionary.formatHelpers;
 
@@ -45,6 +50,31 @@ export function iOSThemeColorsProtocolFormatter(args: FormatterArguments) {
     imports: swiftImports(args.options.imports),
     header: swiftFileHeader(args.file),
     color_tokens: themeColorTokens,
+  });
+}
+
+export function iOSTextStyleFormatter(args: FormatterArguments) {
+  const themeTypographyTokens = _themeTypographyTokens(args.dictionary);
+  const template = compileTemplate('templates/ios/text_style.hbs');
+
+  Handlebars.registerHelper('fontNameRef', (fontFamily, fontWeight) => {
+    const fontWeightName = transformFontWeight(fontWeight);
+    return snakeCase(`${fontFamily} ${fontWeightName}`, {
+      delimiter: '-',
+      transform: part => {
+        return part.charAt(0).toUpperCase() + part.slice(1);
+      },
+    });
+  });
+
+  Handlebars.registerHelper('transformPercentValue', (lineHeight, fontSize) => {
+    return transformPercentValue(lineHeight, fontSize);
+  });
+
+  return template({
+    imports: swiftImports(args.options.imports),
+    header: swiftFileHeader(args.file),
+    typography_tokens: themeTypographyTokens,
   });
 }